+++ src/neuroconv/tools/nwb_helpers/_configuration_models/_base_dataset_io.py
@@ -11,13 +11,14 @@ import zarr
 from hdmf import Container
 from hdmf.data_utils import GenericDataChunkIterator
 from hdmf.utils import get_data_shape
-from pydantic import BaseModel, Field, root_validator
+from pydantic import BaseModel, Field, InstanceOf, PositiveInt, model_validator
 from pynwb import NWBFile
+from typing_extensions import Self
 
 from ...hdmf import SliceableDataChunkIterator
 
 
-def _find_location_in_memory_nwbfile(current_location: str, neurodata_object: Container) -> str:
+def _recursively_find_location_in_memory_nwbfile(current_location: str, neurodata_object: Container) -> str:
     """
     Method for determining the location of a neurodata object within an in-memory NWBFile object.
 
@@ -31,9 +32,18 @@ def _find_location_in_memory_nwbfile(current_location: str, neurodata_object: Co
             if isinstance(parent_field_value, dict) and neurodata_object.name in parent_field_value:
                 return parent_field_name + "/" + neurodata_object.name + "/" + current_location
         return neurodata_object.name + "/" + current_location
-    return _find_location_in_memory_nwbfile(
+    return _recursively_find_location_in_memory_nwbfile(
         current_location=neurodata_object.name + "/" + current_location, neurodata_object=parent
     )
+
+
+def _find_location_in_memory_nwbfile(neurodata_object: Container, field_name: str) -> str:
+    """
+    More readable call for the recursive location finder for a field of a neurodata object in an in-memory NWBFile.
+
+    The recursive method forms from the buttom-up using the initial 'current_location' of the field itself.
+    """
+    return _recursively_find_location_in_memory_nwbfile(current_location=field_name, neurodata_object=neurodata_object)
 
 
 def _infer_dtype_of_list(list_: List[Union[int, float, list]]) -> np.dtype:
@@ -67,104 +77,52 @@ def _infer_dtype(dataset: Union[h5py.Dataset, zarr.Array]) -> np.dtype:
     return data_type
 
 
-class DatasetInfo(BaseModel):
-    """A data model to represent immutable aspects of an object that will become a HDF5 or Zarr dataset on write."""
-
-    # TODO: When using Pydantic v2, replace with
-    # model_config = ConfigDict(allow_mutation=False)
-    class Config:  # noqa: D106
-        allow_mutation = False
-        arbitrary_types_allowed = True
-
-    object_id: str = Field(description="The UUID of the neurodata object containing the dataset.")
-    location_in_file: str = Field(  # TODO: in v2, use init_var=False or assign as a property
-        description="The relative location of the this dataset within the in-memory NWBFile. (e.g. 'acquisition/TestElectricalSeries/data')"
-    )
-    dataset_name: Literal["data", "timestamps"] = Field(description="The reference name of the dataset.")
-    dtype: np.dtype = Field(  # TODO: When using Pydantic v2, replace np.dtype with InstanceOf[np.dtype]
-        description="The data type of elements of this dataset."
-    )
-    full_shape: Tuple[int, ...] = Field(description="The maximum shape of the entire dataset.")
-
-    def __hash__(self):
-        """To allow instances of this class to be used as keys in dictionaries."""
-        return hash((type(self),) + tuple(self.__dict__.values()))
-
-    def __str__(self) -> str:
-        """
-        Not overriding __repr__ as this is intended to render only when wrapped in print().
-
-        Reason being two-fold; a standard `repr` is intended to be slightly more machine-readable / a more basic
-        representation of the true object state. But then also because an iterable of these objects, such as a
-        `List[DataSetInfo]`, would print out the nested representations, which only look good when using the basic
-        `repr` (that is, this fancy string print-out does not look good when nested in another container).
-        """
-        source_size_in_gb = math.prod(self.full_shape) * self.dtype.itemsize / 1e9
-
-        string = (
-            f"\n{self.location_in_file}"
-            f"\n{'-' * len(self.location_in_file)}"
-            f"\n  dtype : {self.dtype}"
-            f"\n  full shape of source array : {self.full_shape}"
-            f"\n  full size of source array : {source_size_in_gb:0.2f} GB"
-        )
-        return string
-
-    def __init__(self, **values):
-        location = values["location_in_file"]
-
-        # For more efficient / explicit reference downstream, instead of reparsing from location multiple times
-        dataset_name = location.split("/")[-1]
-        values.update(dataset_name=dataset_name)
-        super().__init__(**values)
-
-    @classmethod
-    def from_neurodata_object(cls, neurodata_object: Container, field_name: str) -> "DatasetInfo":
-        location_in_file = _find_location_in_memory_nwbfile(
-            current_location=field_name, neurodata_object=neurodata_object
-        )
-        candidate_dataset = getattr(neurodata_object, field_name)
-
-        full_shape = get_data_shape(data=candidate_dataset)
-        dtype = _infer_dtype(dataset=candidate_dataset)
-
-        return cls(
-            object_id=neurodata_object.object_id,
-            object_name=neurodata_object.name,
-            location_in_file=location_in_file,
-            full_shape=full_shape,
-            dtype=dtype,
-        )
-
-
 class DatasetIOConfiguration(BaseModel, ABC):
     """A data model for configuring options about an object that will become a HDF5 or Zarr Dataset in the file."""
 
-    # TODO: When using Pydantic v2, remove
-    class Config:
-        arbitrary_types_allowed = True
+    # Immutable fields about the dataset
+    object_id: str = Field(description="The UUID of the neurodata object containing the dataset.", frozen=True)
+    location_in_file: str = Field(
+        description=(
+            "The location of the this dataset within the in-memory NWBFile relative to the top-level root, "
+            "e.g. 'acquisition/ElectricalSeries/data'."
+        ),
+        frozen=True,
+    )
+    dataset_name: Literal["data", "timestamps"] = Field(description="The reference name of the dataset.", frozen=True)
+    dtype: InstanceOf[np.dtype] = Field(description="The data type of elements of this dataset.", frozen=True)
+    full_shape: Tuple[int, ...] = Field(description="The maximum shape of the entire dataset.", frozen=True)
 
-    dataset_info: DatasetInfo = Field(description="The immutable information about this dataset.")
-    chunk_shape: Tuple[int, ...] = Field(  # When using Pydantic v2.0, specify PositiveInt
+    # User specifiable fields
+    chunk_shape: Union[Tuple[PositiveInt, ...], None] = Field(
         description=(
             "The specified shape to use when chunking the dataset. "
             "For optimized streaming speeds, a total size of around 10 MB is recommended."
-        )
+        ),
     )
-    buffer_shape: Tuple[int, ...] = Field(
+    buffer_shape: Union[Tuple[int, ...], None] = Field(
         description=(
             "The specified shape to use when iteratively loading data into memory while writing the dataset. "
             "For optimized writing speeds and minimal RAM usage, a total size of around 1 GB is recommended."
-        )
+        ),
     )
-    # TODO: When using Pydantic v2, wrap h5py._hl.filters.FilterRefBase and numcodecs.abc.Codec with InstanceOf
-    compression_method: Union[str, h5py._hl.filters.FilterRefBase, numcodecs.abc.Codec, None] = Field(
-        default="gzip",
+    compression_method: Union[
+        str, InstanceOf[h5py._hl.filters.FilterRefBase], InstanceOf[numcodecs.abc.Codec], None
+    ] = Field(
         description="The specified compression method to apply to this dataset. Set to `None` to disable compression.",
     )
     compression_options: Union[Dict[str, Any], None] = Field(
         default=None, description="The optional parameters to use for the specified compression method."
     )
+
+    @abstractmethod
+    def get_data_io_kwargs(self) -> Dict[str, Any]:
+        """
+        Fetch the properly structured dictionary of input arguments.
+
+        Should be passed directly as dynamic keyword arguments (**kwargs) into a H5DataIO or ZarrDataIO.
+        """
+        raise NotImplementedError
 
     def __str__(self) -> str:
         """
@@ -204,12 +162,18 @@ class DatasetIOConfiguration(BaseModel, ABC):
 
         return string
 
-    @root_validator
+    @model_validator(mode="before")
     def validate_all_shapes(cls, values: Dict[str, Any]) -> Dict[str, Any]:
+        location_in_file = values["location_in_file"]
+        dataset_name = values["dataset_name"]
+
+        assert (
+            dataset_name == location_in_file.split("/")[-1]
+        ), f"The `dataset_name` ({dataset_name}) does not match the end of the `location_in_file` ({location_in_file})!"
+
         chunk_shape = values["chunk_shape"]
         buffer_shape = values["buffer_shape"]
         full_shape = values["dataset_info"].full_shape
-        location_in_file = values["dataset_info"].location_in_file  # For more identifiable error messages.
 
         if len(chunk_shape) != len(buffer_shape):
             raise ValueError(
@@ -254,35 +218,51 @@ class DatasetIOConfiguration(BaseModel, ABC):
 
         return values
 
-    @abstractmethod
-    def get_data_io_kwargs(self) -> Dict[str, Any]:
-        """
-        Fetch the properly structured dictionary of input arguments.
-
-        Should be passed directly as dynamic keyword arguments (**kwargs) into a H5DataIO or ZarrDataIO.
-        """
-        raise NotImplementedError
-
     @classmethod
-    def from_neurodata_object(cls, neurodata_object: Container, field_name: str) -> "DatasetIOConfiguration":
-        candidate_dataset = getattr(neurodata_object, field_name)
+    def from_neurodata_object(cls, neurodata_object: Container, dataset_name: Literal["data", "timestamps"]) -> Self:
+        """
+        Construct an instance of a DatasetIOConfiguration for a dataset in a neurodata object in an NWBFile.
 
-        dataset_info = DatasetInfo.from_neurodata_object(neurodata_object=neurodata_object, field_name=field_name)
+        Parameters
+        ----------
+        neurodata_object : hdmf.Container
+            The neurodata object containing the field that will become a dataset when written to disk.
+        dataset_name : "data" or "timestamps"
+            The name of the field that will become a dataset when written to disk.
+            Some neurodata objects can have multiple such fields, such as `pynwb.TimeSeries` which can have both `data`
+            and `timestamps`, each of which can be configured separately.
+        """
+        location_in_file = _find_location_in_memory_nwbfile(neurodata_object=neurodata_object, field_name=dataset_name)
 
-        dtype = dataset_info.dtype
-        full_shape = dataset_info.full_shape
+        candidate_dataset = getattr(neurodata_object, dataset_name)
+        full_shape = get_data_shape(data=candidate_dataset)
+        dtype = _infer_dtype(dataset=candidate_dataset)
 
         if isinstance(candidate_dataset, GenericDataChunkIterator):
             chunk_shape = candidate_dataset.chunk_shape
             buffer_shape = candidate_dataset.buffer_shape
-        elif dtype != "unknown":
+            compression_method = "gzip"
+        elif dtype != np.dtype("object"):
             chunk_shape = SliceableDataChunkIterator.estimate_default_chunk_shape(
                 chunk_mb=10.0, maxshape=full_shape, dtype=np.dtype(dtype)
             )
             buffer_shape = SliceableDataChunkIterator.estimate_default_buffer_shape(
                 buffer_gb=0.5, chunk_shape=chunk_shape, maxshape=full_shape, dtype=np.dtype(dtype)
             )
-        else:
-            pass
+            compression_method = "gzip"
+        elif dtype == np.dtype("object"):  # Unclear what default chunking/compression should be for compound objects
+            chunk_shape = None
+            buffer_shape = None
+            compression_method = None
 
-        return cls(dataset_info=dataset_info, chunk_shape=chunk_shape, buffer_shape=buffer_shape)
+        return cls(
+            object_id=neurodata_object.object_id,
+            object_name=neurodata_object.name,
+            location_in_file=location_in_file,
+            dataset_name=dataset_name,
+            full_shape=full_shape,
+            dtype=dtype,
+            chunk_shape=chunk_shape,
+            buffer_shape=buffer_shape,
+            compression_method=compression_method,
+        )